Java_Lessons-39 29: перегрузка методов 2 https://youtu.be/o37qhweFZBM?t=14m34s Integer

+++ src/OverloadingExmpl.java
@@ -2,7 +2,7 @@ public class OverloadingExmpl {
     public static void main( String[] args ) {
         OverloadingExmpl overloadingExmpl = new OverloadingExmpl();
         short b = 5;
-        overloadingExmpl.method(b);
+        overloadingExmpl.method((int)(b));
     }
 
     void method(Integer i) {
